+++ schema/rss-sled-plan.json
@@ -504,7 +504,7 @@
       }
     },
     "RackNetworkConfigV1": {
-      "description": "Initial network configuration\n\nTODO(AJS): It's unclear if this should be serde renamed to `RackNetworkConfig` I *think* it's useful to have the version be explicit in the name, but I'm open to discussion.",
+      "description": "Initial network configuration",
       "type": "object",
       "required": [
         "bgp",
